Use single [ test for -d condition.
For some reason, zsh 5.0.2 doesn't like the `[[ -d condition ]] &&
success` structure and wants me to use the single [ test instead.

Using the `[[ -d ... ]]` test in a full if statement also works, but
this is closer to the previous version, and less verbose.

This is strange because there are other "inline" tests in this file, but
it only complains about ones that use -d.