Add fzf vim plugin
Note that this requires fzf proper being installed in our vim
environment, but vundle can take care of that for us.

Note also the change to the default command to use `git ls-tree` when in
a git repository; this affects the vim plugin, ctrl-t, **-tab
completion, and pretty much all use of fzf that doesn't use stdin.