Add git-churn and git-diff-churn for churn stats in git repos
git-churn finds edit counts over all time for each file and sorts so the
most-edited files are at the bottom. Piping through tail is common
enough that I might make it part of the command or make another alias
for it.

git-diff-churn finds the same edit counts but filters them through the
list of files that differ between GIT_BASE (default master) and HEAD,
thus giving you a list of files that changed in the current branch and
their respective edit counts.